DEBUG-2334 default token bucket allow? size to 1
The only usage of token bucket rate limiter currently in
our code uses a size of 1 (to perform one action).
Default the size parameter to 1 to simplify the API for
using token bucket on the caller side.

Dynamic instrumentation will also use token bucket and
will also call allow? with the size of 1.